+++ packages/web/docs/src/components/company-logos/index.tsx
@@ -173,3 +173,54 @@ export function KarrotLogo(props: LogoProps) {
     </svg>
   );
 }
+
+export function NacelleLogo(props: LogoProps) {
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 357 91" fill="currentColor" {...props}>
+      <defs>
+        <clipPath id="a">
+          <path fill="none" d="M0 0H91V91H0z" />
+        </clipPath>
+      </defs>
+      <path d="M124 43v23h-9.3V26.7h8.8v6.7h.5c.9-2.2 2.3-3.9 4.3-5.2 2-1.3 4.5-1.9 7.4-1.9s5.1.6 7.1 1.7c2 1.2 3.6 2.8 4.7 5s1.7 4.9 1.7 8v25h-9.3V42.4c0-2.6-.7-4.7-2-6.2s-3.2-2.2-5.6-2.2-3.1.4-4.3 1.1c-1.2.7-2.2 1.7-2.9 3-.7 1.3-1 2.9-1 4.8zM169.9 66.8c-2.5 0-4.7-.4-6.7-1.3s-3.5-2.2-4.7-4c-1.1-1.8-1.7-3.9-1.7-6.5s.4-4 1.2-5.5c.8-1.4 1.9-2.6 3.3-3.5 1.4-.9 3-1.5 4.8-2 1.8-.5 3.6-.8 5.5-1 2.3-.2 4.2-.5 5.6-.6 1.4-.2 2.5-.5 3.1-.9.7-.4 1-1.1 1-2v-.2c0-1.9-.6-3.4-1.7-4.5s-2.8-1.6-4.9-1.6-4.1.5-5.4 1.5c-1.3 1-2.2 2.2-2.7 3.5L158 37c.7-2.4 1.8-4.4 3.4-6 1.6-1.6 3.5-2.8 5.8-3.6 2.3-.8 4.8-1.2 7.5-1.2s3.8.2 5.7.7c1.9.4 3.6 1.2 5.1 2.2 1.6 1 2.8 2.4 3.7 4.1 1 1.7 1.4 3.9 1.4 6.5V66h-8.9v-5.4h-.3c-.6 1.1-1.4 2.1-2.4 3.1-1 .9-2.3 1.7-3.8 2.3-1.5.6-3.3.8-5.3.8zm2.4-6.8c1.9 0 3.5-.4 4.8-1.1 1.4-.8 2.4-1.7 3.1-3 .8-1.2 1.1-2.6 1.1-4v-4.6c-.3.2-.8.5-1.5.7-.7.2-1.4.4-2.3.5-.9.2-1.7.3-2.5.4-.8.1-1.6.2-2.2.3-1.4.2-2.6.5-3.7.9s-2 1-2.6 1.8c-.6.8-.9 1.7-.9 2.9 0 1.7.6 3 1.9 3.9 1.2.9 2.8 1.3 4.8 1.3zM217.1 66.8c-3.9 0-7.3-.9-10.1-2.6-2.8-1.7-5-4.1-6.5-7.1-1.5-3.1-2.2-6.6-2.2-10.5s.8-7.5 2.3-10.6c1.5-3.1 3.7-5.5 6.5-7.2 2.8-1.7 6.1-2.6 10-2.6s6 .6 8.5 1.8c2.5 1.2 4.4 2.8 5.9 4.9 1.5 2.1 2.3 4.6 2.5 7.4h-8.8c-.4-1.9-1.2-3.5-2.6-4.7-1.3-1.3-3.1-1.9-5.3-1.9s-3.6.5-5 1.5-2.5 2.5-3.3 4.3c-.8 1.9-1.2 4.2-1.2 6.8s.4 5 1.2 6.9c.8 1.9 1.9 3.4 3.3 4.4 1.4 1 3.1 1.5 5 1.5s2.6-.3 3.7-.8 2-1.3 2.7-2.3c.7-1 1.2-2.2 1.5-3.6h8.8c-.2 2.8-1 5.2-2.5 7.4-1.4 2.1-3.3 3.8-5.8 5-2.4 1.2-5.3 1.8-8.6 1.8zM258.9 66.8c-3.9 0-7.3-.8-10.2-2.5-2.8-1.7-5-4-6.6-7-1.5-3-2.3-6.6-2.3-10.7s.8-7.6 2.3-10.6c1.6-3.1 3.7-5.5 6.5-7.2 2.8-1.7 6-2.6 9.8-2.6s4.7.4 6.9 1.2c2.2.8 4.1 2 5.7 3.6 1.7 1.6 3 3.7 3.9 6.2 1 2.5 1.4 5.5 1.4 8.9v2.8h-32.2v-6.2h23.3c0-1.8-.4-3.3-1.2-4.7-.8-1.4-1.8-2.5-3.1-3.3-1.3-.8-2.9-1.2-4.7-1.2s-3.6.5-5 1.4c-1.4.9-2.5 2.1-3.4 3.6-.8 1.5-1.2 3.1-1.2 4.8v5.4c0 2.3.4 4.2 1.3 5.9.8 1.6 2 2.9 3.5 3.7 1.5.9 3.3 1.3 5.3 1.3s2.6-.2 3.7-.6c1.1-.4 2-1 2.8-1.7s1.4-1.7 1.8-2.8l8.6 1c-.5 2.3-1.6 4.3-3.1 6-1.5 1.7-3.5 3-5.8 3.9-2.4.9-5.1 1.4-8.1 1.4zM293.5 13.6V66h-9.3V13.6h9.3zM312.2 13.6V66h-9.3V13.6h9.3zM339.1 66.8c-3.9 0-7.3-.8-10.2-2.5-2.8-1.7-5-4-6.6-7-1.5-3-2.3-6.6-2.3-10.7s.8-7.6 2.3-10.6c1.6-3.1 3.7-5.5 6.5-7.2 2.8-1.7 6-2.6 9.8-2.6s4.7.4 6.9 1.2c2.2.8 4.1 2 5.7 3.6 1.7 1.6 3 3.7 3.9 6.2 1 2.5 1.4 5.5 1.4 8.9v2.8h-32.2v-6.2h23.3c0-1.8-.4-3.3-1.2-4.7-.8-1.4-1.8-2.5-3.1-3.3-1.3-.8-2.9-1.2-4.7-1.2s-3.6.5-5 1.4c-1.4.9-2.5 2.1-3.3 3.6-.8 1.5-1.2 3.1-1.2 4.8v5.4c0 2.3.4 4.2 1.3 5.9.8 1.6 2 2.9 3.5 3.7 1.5.9 3.3 1.3 5.3 1.3s2.6-.2 3.7-.6c1.1-.4 2-1 2.8-1.7s1.4-1.7 1.8-2.8l8.6 1c-.5 2.3-1.6 4.3-3.1 6-1.5 1.7-3.5 3-5.8 3.9-2.4.9-5.1 1.4-8.1 1.4z" />
+      <g clipPath="url(#a)">
+        <path d="M20.2 7.7C9.2 15 1.5 27.1.1 40.9l13.4-7.7c3.2-8.5 9.7-15.3 17.8-19.2L20.1 7.6zM18.9 67.1c-4.7-5.9-7.6-13.4-7.6-21.6s0-2.5.2-3.7L0 48.4c.9 14 8.1 26.3 18.9 34V67.1zM51.6 79.7c-1.8.3-3.7.5-5.6.5-7.8 0-14.9-2.5-20.7-6.8v13c6.1 3 12.9 4.7 20.2 4.7s13-1.4 18.7-4l-12.6-7.3zM78.8 57c-3.2 9-10 16.3-18.7 20.2l10.7 6.2C81.8 76 89.5 64 90.9 50.1l-12.1 7zM91 42.6c-.9-14-8.1-26.3-18.9-34v14.1c5.3 6.1 8.6 14.1 8.6 22.8s0 2-.1 3l10.4-6zM45.5 0c-6.7 0-13 1.4-18.7 4l12.9 7.4c2.1-.4 4.2-.6 6.3-.6 7.3 0 14.1 2.3 19.7 6.1V4.7C59.6 1.7 52.8 0 45.5 0z" />
+      </g>
+    </svg>
+  );
+}
+
+export function WealthsimpleLogo(props: LogoProps) {
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 172 26" {...props}>
+      <path
+        fill-rule="evenodd"
+        d="M24.98 2.13v.25a2.7 2.7 0 0 1 1.66 1.14c.38.57.57 1.36.57 2.38 0 1.07-.17 2.15-.5 3.25a22.9 22.9 0 0 1-1.43 3.53l-.68 1.5-3.34-8.1a7.87 7.87 0 0 1-.55-2.13c0-.88.48-1.4 1.46-1.57v-.25H12.2v.25c.38.07.71.19.99.34.27.16.54.43.79.82.24.39.52.96.83 1.71l1.6 3.73-2.82 5.62-4.2-9.76a2.56 2.56 0 0 1-.26-.6 1.88 1.88 0 0 1-.07-.49c0-.29.1-.56.3-.8.2-.24.5-.43.88-.58v-.25H.08v.25c.8.24 1.53.7 2.17 1.36.65.66 1.28 1.68 1.9 3.04L10.6 21.2h.28l5.81-11.6 4.98 11.59h.28l5.92-12.82a21.17 21.17 0 0 1 2.38-4.13 6.25 6.25 0 0 1 2.28-1.88v-.25h-7.55v.01Z"
+        clip-rule="evenodd"
+      />
+      <path d="M41.26 15.09c-.65 1.64-2.07 2.46-4.27 2.46a4.2 4.2 0 0 1-3.24-1.28c-.76-.82-1.16-2.05-1.2-3.67h8.9v-.17c0-1.14-.3-2.14-.88-2.99a6 6 0 0 0-2.36-2c-1-.47-2.1-.71-3.34-.71a7.4 7.4 0 0 0-3.64.91 6.98 6.98 0 0 0-3.65 6.34c0 1.35.3 2.55.9 3.59A6.47 6.47 0 0 0 31 20.02c1.07.6 2.3.88 3.67.88 1.84 0 3.34-.47 4.52-1.43a7.26 7.26 0 0 0 2.37-4.3l-.3-.08Zm-8.13-6.65c.39-.88.95-1.33 1.7-1.33a1.4 1.4 0 0 1 1.24.7c.13.2.24.44.34.7.09.27.16.63.2 1.08.05.46.07 1.06.07 1.82v.69h-4.12c0-.84.05-1.54.14-2.1.1-.63.25-1.15.43-1.56Zm23.94 10.25a.6.6 0 0 1-.32.15l-.31.02a.73.73 0 0 1-.7-.4c-.14-.27-.21-.8-.21-1.6v-6.48c0-1.3-.28-2.23-.83-2.82-.55-.6-1.43-.89-2.64-.89a16.25 16.25 0 0 0-6.75 1.6 7.62 7.62 0 0 0-2.33 1.59 2.67 2.67 0 0 0-.79 1.82c0 .58.2 1.03.58 1.37.39.35.9.52 1.52.52a2.5 2.5 0 0 0 1.81-.8 3 3 0 0 0 .8-2.22v-.47c0-.85.2-1.49.57-1.93.37-.43.84-.65 1.39-.65.44 0 .8.11 1.09.32.28.21.5.58.63 1.1.14.51.2 1.23.2 2.16v1.41c-1.39.43-2.62.84-3.7 1.26-1.22.47-2.24.93-3.04 1.4-1.27.71-1.9 1.64-1.9 2.76 0 .89.35 1.6 1.06 2.13.7.54 1.63.8 2.77.8a6.4 6.4 0 0 0 2.93-.64c.7-.35 1.35-.9 1.95-1.67.14.78.4 1.34.8 1.7a2.6 2.6 0 0 0 1.83.64 4.3 4.3 0 0 0 2.1-.52 5.01 5.01 0 0 0 1.65-1.47l-.16-.19Zm-7.88.25c-1.05 0-1.57-.72-1.57-2.16 0-.96.29-1.75.86-2.37.48-.53 1.26-1 2.31-1.42v4c0 .62-.15 1.1-.44 1.43-.3.35-.68.52-1.16.52Z" />
+      <path
+        fill-rule="evenodd"
+        d="M57.07 20.18c.53-.04.92-.24 1.17-.62.25-.38.37-1.14.37-2.29V5.95c0-1.3-.17-2.22-.52-2.76a1.67 1.67 0 0 0-1.49-.82v-.3A66.02 66.02 0 0 0 63.35.19v17.64c0 .9.1 1.5.29 1.82.2.32.53.5 1 .53v.3h-7.57v-.3Zm9.97-.05c-.55-.52-.83-1.4-.83-2.63V8.22h-2.04L64.12 8c.72-.33 1.48-.81 2.3-1.44a22.02 22.02 0 0 0 4.42-4.54l.25.09v4.95h3.08l-.66 1.17H71.1v8.41c0 .82.12 1.4.36 1.75.24.35.59.53 1.05.53.22 0 .44-.03.67-.09.23-.05.44-.15.62-.28l.16.23c-.8.68-1.57 1.2-2.28 1.57-.72.37-1.47.56-2.26.56a3.37 3.37 0 0 1-2.37-.78Z"
+        clip-rule="evenodd"
+      />
+      <path
+        fill-rule="evenodd"
+        d="M88.76 19.6a5.28 5.28 0 0 1-.26-2.05V10.5c0-2.54-1.03-3.82-3.08-3.82-.6 0-1.13.07-1.62.22a6.2 6.2 0 0 0-1.6.8c-.54.37-1.2.88-2 1.53V.2a67.7 67.7 0 0 1-6.74 1.88v.3c.64 0 1.14.28 1.49.82.35.55.52 1.47.52 2.76V17c0 1.23-.13 2.06-.4 2.49-.26.43-.68.65-1.25.7v.3h7.66v-.3c-.46-.05-.79-.24-.98-.59a4.85 4.85 0 0 1-.29-2.05V9.71c.41-.34.78-.6 1.1-.74.37-.17.7-.25 1-.25.47 0 .84.2 1.11.6.27.4.4 1.08.4 2.06v6.14c0 .72-.04 1.27-.11 1.64-.07.36-.2.62-.37.76-.18.13-.44.22-.79.26v.3h7.16v-.3c-.46-.03-.77-.23-.95-.58Zm6.08 1.1-1.8-.35a7.49 7.49 0 0 0-1.43-.17 2.46 2.46 0 0 0-1.16.28v-6.2l.28-.06a29.62 29.62 0 0 0 2.64 3.64c.8.9 1.53 1.54 2.2 1.91.69.38 1.36.57 2.02.57.64 0 1.11-.14 1.4-.43.3-.29.44-.67.44-1.15 0-.59-.2-1.06-.59-1.4a7.4 7.4 0 0 0-2.1-1.14l-1.96-.78a9.85 9.85 0 0 1-3.3-1.91 3.3 3.3 0 0 1-.94-2.4c0-.84.21-1.58.65-2.25a4.5 4.5 0 0 1 1.86-1.59c.8-.4 1.77-.6 2.89-.6.59 0 1.17.05 1.75.14.58.1 1.1.18 1.56.27.46.08.8.12 1.04.12a1.63 1.63 0 0 0 .75-.14l.38 5.07-.25.06a16.96 16.96 0 0 0-3.03-3.8A4.16 4.16 0 0 0 95.5 7.2c-.55 0-1 .14-1.35.42-.35.3-.52.68-.52 1.18 0 .54.17.98.53 1.33.36.35.96.67 1.8.94l2.32.78a7.99 7.99 0 0 1 3.13 1.67c.63.64.95 1.45.95 2.42a4.5 4.5 0 0 1-1.53 3.6 5.98 5.98 0 0 1-4.06 1.33 11.3 11.3 0 0 1-1.93-.18Zm11.34-14.89a2.85 2.85 0 0 0 2.87-2.88c0-.75-.26-1.42-.79-1.99a2.69 2.69 0 0 0-2.08-.86c-.5 0-.96.14-1.39.4-.43.27-.78.62-1.05 1.06a2.77 2.77 0 0 0 .45 3.43c.58.56 1.24.84 2 .84Z"
+        clip-rule="evenodd"
+      />
+      <path
+        fill-rule="evenodd"
+        d="M133.69 19.66c-.2-.32-.29-.92-.29-1.83V10.5c0-2.54-1.03-3.82-3.08-3.82-.57 0-1.13.09-1.67.27-.54.17-1.1.45-1.7.84-.52.34-1.1.79-1.76 1.33a2.6 2.6 0 0 0-.85-1.65 3.25 3.25 0 0 0-2.26-.79c-.9 0-1.79.21-2.67.64-.8.38-1.59.96-2.4 1.73V6.67a55.75 55.75 0 0 1-4.43 1.5l-1.95.55V9c.6.05 1.03.3 1.28.74s.37 1.3.37 2.58v5.5c0 .9-.16 1.5-.5 1.85-.33.34-.73.51-1.21.51-.57 0-1-.17-1.28-.51-.29-.34-.43-.96-.43-1.84V6.6c-1.02.41-2.1.78-3.25 1.13-1.14.34-2.3.66-3.5.95v.28c.74.05 1.26.35 1.57.9.32.54.47 1.45.47 2.7V17c0 1.18-.13 2-.39 2.45-.25.45-.67.7-1.26.73v.3h15.78v-.3a1.22 1.22 0 0 1-.96-.6 3.8 3.8 0 0 1-.3-1.78V9.6c.33-.28.63-.5.9-.62a2.5 2.5 0 0 1 1.03-.23c.53 0 .93.18 1.19.56.25.38.38 1.07.38 2.07v6.42c0 .9-.1 1.51-.29 1.82-.19.32-.5.5-.95.56v.3h7.22v-.3c-.48-.06-.81-.23-1-.53-.18-.3-.27-.91-.27-1.85V9.6c.35-.3.66-.53.94-.7a2 2 0 0 1 1.05-.32c.47 0 .85.2 1.11.62.27.42.4 1.14.4 2.18v6.42c0 .96-.08 1.58-.25 1.87-.16.29-.5.45-1.01.51v.3h7.3v-.3c-.53-.03-.89-.2-1.08-.52Z"
+        clip-rule="evenodd"
+      />
+      <path d="M149.53 10.58a5.25 5.25 0 0 0-1-2.14 4.51 4.51 0 0 0-3.64-1.8 5.3 5.3 0 0 0-2.19.46c-.59.27-1.21.67-1.86 1.2V6.65a79.25 79.25 0 0 1-6.72 2.07v.25c.51.05.91.18 1.2.39.28.2.49.53.62 1 .13.46.19 1.1.19 1.93v9.36c0 1-.09 1.76-.25 2.3-.16.53-.41.9-.74 1.12a2.8 2.8 0 0 1-1.24.4v.3h9v-.3c-.77-.07-1.3-.3-1.6-.69-.31-.39-.46-1.12-.46-2.19v-2.54c.36.25.7.44 1.05.55a6.27 6.27 0 0 0 4.05-.26 5.7 5.7 0 0 0 2.01-1.43 7 7 0 0 0 1.4-2.38 9 9 0 0 0 .5-3.07c0-1.06-.1-2.02-.32-2.88Zm-5.47 8.03c-.51.94-1.15 1.4-1.9 1.4a1.9 1.9 0 0 1-.85-.17 2.6 2.6 0 0 1-.47-.32V8.86c.61-.48 1.14-.72 1.57-.72.7 0 1.27.47 1.73 1.4.46.93.7 2.4.7 4.39 0 2.17-.27 3.73-.78 4.68Z" />
+      <path
+        fill-rule="evenodd"
+        d="M149.7 20.18c.54-.04.93-.24 1.18-.62.25-.38.37-1.14.37-2.29V5.95c0-1.3-.17-2.22-.52-2.76a1.67 1.67 0 0 0-1.49-.82v-.3a93.6 93.6 0 0 0 3.44-.87c1.18-.33 2.28-.66 3.3-1.01v17.64c0 .9.1 1.5.3 1.82.19.32.53.5 1 .53v.3h-7.57v-.3Z"
+        clip-rule="evenodd"
+      />
+      <path d="M170.83 15.09c-.64 1.64-2.06 2.46-4.27 2.46a4.2 4.2 0 0 1-3.23-1.28c-.76-.82-1.16-2.05-1.2-3.67h8.9v-.17c0-1.14-.3-2.14-.88-2.99a6 6 0 0 0-2.37-2 7.67 7.67 0 0 0-3.33-.71c-1.32 0-2.54.3-3.65.91a6.98 6.98 0 0 0-3.65 6.34c0 1.35.3 2.55.91 3.59a6.47 6.47 0 0 0 2.51 2.45c1.07.6 2.29.88 3.67.88 1.83 0 3.34-.47 4.51-1.43a7.26 7.26 0 0 0 2.37-4.3l-.29-.08Zm-8.12-6.65c.38-.88.95-1.33 1.7-1.33.26 0 .5.06.7.17.22.11.4.29.54.53.13.2.24.44.33.7.1.27.17.63.21 1.08.05.46.07 1.06.07 1.82v.69h-4.13c.01-.84.06-1.54.15-2.1.1-.63.25-1.15.43-1.56Z" />
+    </svg>
+  );
+}
